api_mapas/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
branca==0.8.1
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
dnspython==2.8.0
executing==2.2.0
fastapi==0.117.1
folium==0.20.0
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
motor==3.7.1
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
parso==0.8.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.15.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310; platform_system=="Windows"
pyzmq==27.0.0
requests==2.32.5
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.48.0
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
xyzservices==2025.4.0
uvicorn==0.36.0
python-dotenv==1.1.1